+++ apps/backend/src/lib/tokens.tsx
@@ -3,10 +3,11 @@ import { KnownErrors } from '@stackframe/stack-shared';
 import { yupNumber, yupObject, yupString } from "@stackframe/stack-shared/dist/schema-fields";
 import { generateSecureRandomString } from '@stackframe/stack-shared/dist/utils/crypto';
 import { getEnvVariable } from '@stackframe/stack-shared/dist/utils/env';
-import { decryptJWE, encryptJWE, signJWT, verifyJWT } from '@stackframe/stack-shared/dist/utils/jwt';
+import { legacySignGlobalJWT, legacyVerifyGlobalJWT, signJWT, verifyJWT } from '@stackframe/stack-shared/dist/utils/jwt';
+import { Result } from '@stackframe/stack-shared/dist/utils/results';
+import * as jose from 'jose';
 import { JOSEError, JWTExpired } from 'jose/errors';
 import { SystemEventTypes, logEvent } from './events';
-import { Result } from '@stackframe/stack-shared/dist/utils/results';
 
 export const authorizationHeaderSchema = yupString().matches(/^StackSession [^ ]+$/);
 
@@ -39,7 +40,16 @@ const jwtIssuer = "https://access-token.jwt-signature.stack-auth.com";
 export async function decodeAccessToken(accessToken: string) {
   let payload;
   try {
-    payload = await verifyJWT(jwtIssuer, accessToken);
+    const decoded = jose.decodeJwt(accessToken);
+
+    if (!decoded.aud) {
+      payload = await legacyVerifyGlobalJWT(jwtIssuer, accessToken);
+    } else {
+      payload = await verifyJWT({
+        issuer: jwtIssuer,
+        jwt: accessToken,
+      });
+    }
   } catch (error) {
     if (error instanceof JWTExpired) {
       return Result.error(new KnownErrors.AccessTokenExpired());
@@ -50,7 +60,7 @@ export async function decodeAccessToken(accessToken: string) {
   }
 
   const result = await accessTokenSchema.validate({
-    projectId: payload.projectId,
+    projectId: payload.aud || payload.projectId,
     userId: payload.sub,
     refreshTokenId: payload.refreshTokenId,
     exp: payload.exp,
@@ -59,41 +69,50 @@ export async function decodeAccessToken(accessToken: string) {
   return Result.ok(result);
 }
 
-export async function generateAccessToken({
-  projectId,
-  userId,
-}: {
+export async function generateAccessToken(options: {
   projectId: string,
+  useLegacyGlobalJWT: boolean,
   userId: string,
 }) {
-  await logEvent([SystemEventTypes.UserActivity], { projectId, userId });
+  await logEvent([SystemEventTypes.UserActivity], { projectId: options.projectId, userId: options.userId });
 
-  return await signJWT(jwtIssuer, { projectId, sub: userId }, getEnvVariable("STACK_ACCESS_TOKEN_EXPIRATION_TIME", "10min"));
+  if (options.useLegacyGlobalJWT) {
+    return await legacySignGlobalJWT(
+      jwtIssuer,
+      { projectId: options.projectId, sub: options.userId },
+      getEnvVariable("STACK_ACCESS_TOKEN_EXPIRATION_TIME", "10min")
+    );
+  } else {
+    return await signJWT({
+      issuer: jwtIssuer,
+      audience: options.projectId,
+      payload: { sub: options.userId },
+      expirationTime: getEnvVariable("STACK_ACCESS_TOKEN_EXPIRATION_TIME", "10min"),
+    });
+  }
 }
 
-export async function createAuthTokens({
-  projectId,
-  projectUserId,
-  expiresAt,
-}: {
+export async function createAuthTokens(options: {
   projectId: string,
   projectUserId: string,
+  useLegacyGlobalJWT: boolean,
   expiresAt?: Date,
 }) {
-  expiresAt ??= new Date(Date.now() + 1000 * 60 * 60 * 24 * 365);
+  options.expiresAt ??= new Date(Date.now() + 1000 * 60 * 60 * 24 * 365);
 
   const refreshToken = generateSecureRandomString();
   const accessToken = await generateAccessToken({
-    projectId,
-    userId: projectUserId,
+    projectId: options.projectId,
+    userId: options.projectUserId,
+    useLegacyGlobalJWT: options.useLegacyGlobalJWT,
   });
 
   await prismaClient.projectUserRefreshToken.create({
     data: {
-      projectId,
-      projectUserId,
+      projectId: options.projectId,
+      projectUserId: options.projectUserId,
       refreshToken: refreshToken,
-      expiresAt,
+      expiresAt: options.expiresAt,
     },
   });