+++ packages/stack/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @stackframe/stack
 
+## 2.6.6
+
+### Patch Changes
+
+- Bugfixes
+  - @stackframe/stack-sc@2.6.6
+  - @stackframe/stack-shared@2.6.6
+  - @stackframe/stack-ui@2.6.6
+
+## 2.6.5
+
+### Patch Changes
+
+- Minor improvements
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.5
+  - @stackframe/stack-sc@2.6.5
+  - @stackframe/stack-ui@2.6.5
+
+## 2.6.4
+
+### Patch Changes
+
+- fixed small problems
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.4
+  - @stackframe/stack-sc@2.6.4
+  - @stackframe/stack-ui@2.6.4
+
+## 2.6.3
+
+### Patch Changes
+
+- Bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.3
+  - @stackframe/stack-ui@2.6.3
+  - @stackframe/stack-sc@2.6.3
+
+## 2.6.2
+
+### Patch Changes
+
+- Several bugfixes & typos
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.2
+  - @stackframe/stack-ui@2.6.2
+  - @stackframe/stack-sc@2.6.2
+
 ## 2.6.1
 
 ### Patch Changes